Refactor controllers into package controller
