+++ android-app/app/src/main/java/tallerii/stories/controller/LoginController.java
@@ -1,10 +1,11 @@
-package tallerii.stories;
+package tallerii.stories.controller;
 
 import android.util.Log;
 import android.widget.Toast;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
+import tallerii.stories.LoginActivity;
 import tallerii.stories.network.AdapterApplicationApiRest;
 import tallerii.stories.network.EndpointsApplicationApiRest;
 import tallerii.stories.network.apimodels.User;
