+++ android-app/app/src/test/java/tallerii/stories/LoginControllerTest.java
@@ -11,6 +11,7 @@ import org.junit.runners.JUnit4;
 
 import java.io.IOException;
 
+import tallerii.stories.controller.LoginController;
 import tallerii.stories.network.ConstantsApplicationApiRest;
 
 import static org.mockito.ArgumentMatchers.anyString;
